uagrm_bot_backend/requirements.txt
django
djangorestframework
drf-yasg
python-docx==0.8.11
celery==5.6.1
redis==7.1.0
python-dotenv
psycopg2-binary
openai
pgvector
llama-index
llama-parse
nest_asyncio
django-cors-headers
